Renamed a few packages to internal in the sensebox binding. (#2738)
See openhab/static-code-analysis#100
