+++ addons/binding/org.openhab.binding.sensebox/src/main/java/org/openhab/binding/sensebox/internal/model/SenseBoxLoc.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.sensebox.model;
+package org.openhab.binding.sensebox.internal.model;
 
 import com.google.gson.annotations.SerializedName;
